+++ cypress/support/constants/constants.js
@@ -387,6 +387,7 @@ module.exports = {
     SENDVOICECOMMAND: 'fcs.sendVoiceCommand',
     NOTIFY_FIREBOLT_VERSION: 'fcs.notifyFireboltVersion',
     CLEAR_ALL_LISTENERS: 'fcs.clearAllListeners',
+    FINDLOGPATTERN: 'fcs.findLogPattern',
   },
   RESPONSE: 'Response: ',
   RESPONSE_INVALID_RESPONSE_LOG:
@@ -542,10 +543,12 @@ module.exports = {
   DISMISSED: 'dismissed',
   CLOSED: 'closed',
   UNLOADED: 'unloaded',
+  STREAMING: 'streaming',
   ACTIONTYPE: {
     DISMISS_APP: 'dismissApp',
     CLOSE_APP: 'closeApp',
     UNLOAD_APP: 'unloadApp',
+    PLAYBACK: 'playback',
   },
   SUPPORTS_PLATFORM_COMMUNICATION: 'supportsPlatformCommunication',
   FCS_EXIT_CODE: {