+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.13.0-dev.7](https://github.com/rdkcentral/firebolt-certification-suite/compare/v1.13.0-dev.6...v1.13.0-dev.7) (2025-06-25)
+
+
+### Features
+
+* Continue watching implementation-new ([#440](https://github.com/rdkcentral/firebolt-certification-suite/issues/440)) ([4ef4d01](https://github.com/rdkcentral/firebolt-certification-suite/commit/4ef4d015136c84acf39cf1b736a66c0d0c9872cd))
+
+# [1.13.0-dev.6](https://github.com/rdkcentral/firebolt-certification-suite/compare/v1.13.0-dev.5...v1.13.0-dev.6) (2025-06-24)
+
+
+### Features
+
+* Added capture screenshot command ([#435](https://github.com/rdkcentral/firebolt-certification-suite/issues/435)) ([fad9297](https://github.com/rdkcentral/firebolt-certification-suite/commit/fad9297fc130f512daca29ba0124e26475c1a4c8))
+
 # [1.13.0-dev.5](https://github.com/rdkcentral/firebolt-certification-suite/compare/v1.13.0-dev.4...v1.13.0-dev.5) (2025-06-18)